pyproject.toml
[project]
name = "archivey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "davi davi", email = "gh@davitf.com" }
]
requires-python = ">=3.10"
dependencies = [
    "backports-strenum>=1.3.1 ; python_full_version < '3.11'",
    "tqdm>=4.67.1",
]

[project.scripts]
archivey = "archivey:main"

[project.optional-dependencies]
optional = [
    "lz4>=4.4.4",
    "py7zr>=0.22.0",
    "rarfile>=4.2",
    "zstandard>=0.23.0",
]
docs = [
    "sphinx >= 7.0",
    "furo",
    "sphinx-autodoc-typehints >= 2.0",
    "myst-parser >= 2.0",
    "sphinx-copybutton >= 0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "hatch>=1.14.1",
    "pytest>=8.3.5",
    "ruff>=0.11.11",
]

[tool.pytest.ini_options]
# pythonpath = ["src", "tests"]
testpaths = ["tests"]
addopts = "--log-level=DEBUG"

[tool.hatch.envs.default.scripts]
lint = "ruff check --fix . && ruff format . && npx pyright ."
test = "pytest"
docs = "sphinx-build -b html docs docs/_build"

[tool.ruff]
target-version = "py310"
lint.extend-select = ["I"]  # sort imports
